+++ spring-cloud-starter-tencent-polaris-contract/src/main/java/com/tencent/cloud/polaris/contract/tsf/TsfContractPropertiesBootstrapConfiguration.java
@@ -17,7 +17,7 @@
 
 package com.tencent.cloud.polaris.contract.tsf;
 
-import com.tencent.cloud.polaris.context.tsf.ConditionalOnTsfEnabled;
+import com.tencent.cloud.common.tsf.ConditionalOnTsfEnabled;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
